Split DiveMathSupport (and the flint subclass) into its own file.  Split DiveMesh (and subclasses) into its own file too.
